configure clap to accept only values >=2 for option --num-tasks

+++ src/cli.rs
@@ -16,6 +16,6 @@ pub struct Args {
     #[arg(long, default_value_t = String::from("firewall/firewall.txt"))]
     pub firewall_path: String,
     /// Number of asynchronous tasks to use (AKA coroutines)
-    #[arg(long, default_value_t = 2)]
-    pub num_tasks: usize,
+    #[arg(long, default_value_t = 2, value_parser=clap::value_parser!(u8).range(2..))]
+    pub num_tasks: u8,
 }
